Rename deployTestingPriceFeed
